Use the Compile Examples workflow for the deltas report
Using a scheduled workflow for the size deltas report is not appropriate in private repositories because it uses a lot
of minutes and, unlike public repositories, private repositories have a limited allotment of free GitHub Actions
minutes.

The `report-size-deltas` job is configured for use in private repositories which don't have the "Send write tokens to
workflows from fork pull requests" setting enabled. This means that PRs submitted from forks won't get size deltas
reports. If reports on PRs from forks are wanted, the repository can be configured to provide them with write tokens and
the workflow adjusted so the job will run on PRs from forks

Once the repository is made public, the workflow should be reconfigured to provide size deltas reports on PRs from
forks.

More information:
https://github.com/arduino/report-size-deltas#sketches-reports-source