Add manual trigger events to sketch compile workflow
The `workflow_dispatch` and `repository_dispatch` events allow manual triggering of the workflow.

These can be useful to immediately check for impacts of changes to external resources. They do no harm when they are not
needed, so there is no reason not to have them in place to be ready for when they might be needed.