Always upload sketch compilation report artifact
The `arduino/compile-sketches` action produces a report of the results of the compilations. The primary usage of this
report is to generate the size deltas comment that is made to pull requests. Although that can only be done for pull
requests, it might be that the compilation report is found useful for other purposes, in which case it is just as
relevant when triggered by non-PR events.

Always uploading the compilation report also makes the workflow a little more simple.