Fail sketch compile workflow if report not found
The "arduino/compile-sketches` GitHub Actions action generates a report of the results of the compilations as a JSON file.
This file is uploaded as a workflow run artifact, which is later consumed by the `arduino/report-size-deltas` action.

By default, the `actions/upload-artifact` used for the upload step does not require the target file to exist. In this
usage, the report file should always exist. If not, it indicates something is seriously wrong and the workflow run should
fail to bring this to our attention rather than silently ignoring it.