Make size deltas report even when some compilations failed
The GitHub Actions workflow job that does a "smoke test" compilation of the library example sketches is a matrix, running
a parallel job for each of the boards in the array. It may be that one job passes and another fails. In this case, there
is memory usage change data available for the board that passed compilation. Previously the size deltas report only
happened if all compilations passed.

It may be useful for the contributor to get the size deltas feedback in addition to the compilation results even when one
of the boards did not pass compilation so that they may have as much of the available information as possible to work with.