Use major version refs of `arduino/report-size-deltas` action
The `arduino/report-size-deltas` GitHub Actions action is used to provide feedback about the impact of a proposed change
on sketch memory usage. Previously, due to the lack of a release, the development version of the action was used. Using
release versions provides a more stable CI system for the project.

Use of the major version ref will cause the workflow to benefit from ongoing development to the actions up until such
time as a new major release of an action is made, at which time we would need to evaluate whether any changes to the
workflow are required by the breaking change that triggered the major release before updating the major ref
(e.g., uses: `arduino/report-size-deltas@v2`).