Add descriptive name to size deltas report step
The `name` key of a GitHub Actions workflow step is used to define the name used to identify that step to humans. When
one is not defined, a more cryptic name is generated from the machine readable content of the step. A descriptive name
will make it easier to understand the workflow run logs.