Add CI workflow to do Arduino project-specific linting
On every push, pull request, and periodically, run Arduino Lint to check for common problems not related to the project
code.

The URLs in the workflow status badge are configured for the final location of the library, so they are expected to be broken while it is in the temporary development location.