Use major version ref of `arduino/compile-sketches` action (#24)
The `arduino/compile-sketches` GitHub Actions action is used by the CI system to compile the library's example sketches
for each of the supported boards in order to provide a "smoke test". Previously, due to the use of an older modified
variant of Arduino CLI, a specific ref of the action had to be used. The variant of Arduino CLI has since been brought up to date, allowing it to be used with the current version of the action.

Use of the major version ref will cause the workflow to use a stable version of the action while also benefit from
ongoing development to the action up until such time as a new major release of an action is made, at which time we would
need to evaluate whether any changes to the workflow are required by the breaking change that triggered the major release
before updating the major ref (e.g., uses: `arduino/compile-sketches@v2`).