Configure size deltas report for public repository
This repository uses the "arduino/report-size-deltas" GitHub Actions action to cause a report to be commented to pull
request threads of the change in the memory usage of the example sketches caused by the changes proposed in the pull
request.

This system was set up during the initial development of the library, when the repository was private. The configuration
was specially tailored for use in a private repository. Since that time, the repository has been made public, which has
implications for the configuration of this report system.

In order to make this comment, the GitHub Actions workflow must have write permissions in the repository. The "Compile
Examples" workflow will not have these permissions when it is triggered by a pull request from a fork. This causes the
"report-size-deltas" job of the workflow to fail. For this reason, there a different configuration of the report system
is used in public repositories. The alternative configuration uses a dedicated workflow to produce the report. That
workflow is triggered on a schedule rather than by the pull request event. That gives it the required write permissions
in the repository.