Fix typos in documentation
The "codespell" spellchecker tool is used to automatically detect commonly misspelled words in the files of this
project.

The misspelled words dictionary was expanded in the latest release of codespell, which resulted in the detection of a
misspelled word in the project's comments:

> Error: ./docs/threadsafe-serial.md:51: pre-emptive ==> preemptive
> Codespell found one or more problems

The typo is hereby corrected, which will restore the spell check to a passing state.

---

I supplemented the correction of the automatically detected typo with a general review of the documentation content,
fixing the additional typos I identified during that review.